refactor screen structure. now is time for design portion.
