+++ Frontend/app/lesson.tsx
@@ -1,4 +1,4 @@
-import Reat from 'react';
+import React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
 
 export default function LessonScreen() {
